requirements.txt
anyio==4.4.0
certifi==2024.6.2
dnspython==2.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pymongo==4.7.3
python-dotenv==1.0.1
python-telegram-bot==21.3
sniffio==1.3.1
